requirements.txt
pyglet==1.5.21
moderngl==5.6.4
moderngl-window==2.4.0
pyopengl==3.1.5
pyglm==2.4.0
imgui==1.3.0
